Arreglo el pool de conexiones

+++ revistas-digitales-api/src/main/java/com/ipc2/revistas/digitales/api/controladores/AnuncioController.java
@@ -59,7 +59,7 @@ public class AnuncioController {
             @FormDataParam("urlVideo") String urlVideo,
             @FormDataParam("duracion") Integer duracion,
             @FormDataParam("fechaInicio") String fechaInicio,
-            @FormDataParam("totalAPagar") Integer totalAPagar) {
+            @FormDataParam("totalAPagar") double totalAPagar) {
 
         boolean anuncioComprado = anuncioService.comprarAnuncio(anuciante, tipoAnuncio, contenidoTexto, imagenInputStream, urlVideo, duracion, fechaInicio, totalAPagar);
